+++ TheChessProject/src/no/pdigre/chess/engine/base/NodeUtil.java
@@ -5,32 +5,32 @@ import no.pdigre.chess.engine.base.IConst.BITS;
 
 public class NodeUtil {
 
-    public static int[] filterFrom(int[] moves, int from) {
+    public static long[] filterFrom(long[] moves, int from) {
         int length = 0;
-        int[] array = new int[moves.length];
+        long[] array = new long[moves.length];
         for (int i = 0; i < moves.length; i++) {
-            int bitmap = moves[i];
+        	long bitmap = moves[i];
             if (BITS.getFrom(bitmap) == from) {
                 array[length] = bitmap;
                 length++;
             }
         }
-        int[] ret = new int[length];
+        long[] ret = new long[length];
         System.arraycopy(array, 0, ret, 0, length);
         return ret;
     }
 
-    public static int[] filterTo(int[] moves, int to) {
+    public static long[] filterTo(long[] moves, int to) {
         int length = 0;
-        int[] array = new int[moves.length];
+        long[] array = new long[moves.length];
         for (int i = 0; i < moves.length; i++) {
-            int bitmap = moves[i];
+        	long bitmap = moves[i];
             if (BITS.getTo(bitmap) == to) {
                 array[length] = bitmap;
                 length++;
             }
         }
-        int[] ret = new int[length];
+        long[] ret = new long[length];
         System.arraycopy(array, 0, ret, 0, length);
         return ret;
     }
