+++ TheChessProject/src/no/pdigre/chess/test/util/CountFull.java
@@ -30,7 +30,7 @@ public class CountFull extends RecursiveTask<Counter[]> {
 
     protected void count(IPosition pos) {
         counters[0].moves++;
-        int bitmap = pos.getBitmap();
+        long bitmap = pos.getBitmap();
         if (BITS.isCapture(bitmap)) {
             counters[0].captures++;
             if (BITS.isEnpassant(bitmap))
